+++ src/pages/change-mot-passe.tsx
@@ -7,7 +7,7 @@ import { useBreadcrumb } from "../frontend/ui/commun/hooks/useBreadcrumb";
 
 export default function MotPasseOublie() {
   const { wording } = useDependencies();
-  const { setIsInfoPage } = useContext(BackToSearchContext) as BackToSearchContextValue;
+  const backToSearchContext = useContext(BackToSearchContext) as BackToSearchContextValue;
 
   useBreadcrumb([
     {
@@ -17,9 +17,11 @@ export default function MotPasseOublie() {
   ]);
 
   useEffect(() => {
-    setIsInfoPage(false);
-    localStorage.clear();
-  }, []);
+    if (backToSearchContext) {
+      backToSearchContext.setIsInfoPage(false);
+      localStorage.clear();
+    }
+  }, [backToSearchContext])
 
   return <ChangePwdPage />;
 }
