+++ src/pages/connexion.tsx
@@ -5,14 +5,16 @@ import { useBreadcrumb } from "../frontend/ui/commun/hooks/useBreadcrumb";
 import { PageDeConnexion } from "../frontend/ui/login/PageDeConnexion";
 
 export default function PageDAccueil() {
-  const { setIsInfoPage } = useContext(BackToSearchContext) as BackToSearchContextValue;
+  const backToSearchContext = useContext(BackToSearchContext) as BackToSearchContextValue;
 
   useBreadcrumb([]);
 
   useEffect(() => {
-    setIsInfoPage(false);
-    localStorage.clear();
-  }, []);
+    if (backToSearchContext) {
+      backToSearchContext.setIsInfoPage(false);
+      localStorage.clear();
+    }
+  }, [backToSearchContext])
 
   return <PageDeConnexion />;
 }
