+++ src/pages/eco-conception.tsx
@@ -8,7 +8,7 @@ import { PageEcoConception } from "../frontend/ui/eco-conception/PageEcoConcepti
 
 export default function EcoConception() {
   const { wording } = useDependencies();
-  const { setIsInfoPage } = useContext(BackToSearchContext) as BackToSearchContextValue;
+  const backToSearchContext = useContext(BackToSearchContext) as BackToSearchContextValue;
 
   useBreadcrumb([
     {
@@ -18,9 +18,11 @@ export default function EcoConception() {
   ]);
 
   useEffect(() => {
-    setIsInfoPage(false);
-    localStorage.clear();
-  }, []);
+    if (backToSearchContext) {
+      backToSearchContext.setIsInfoPage(false);
+      localStorage.clear();
+    }
+  }, [backToSearchContext])
 
   return (
     <main className="fr-container">
