+++ src/pages/registration.tsx
@@ -5,13 +5,15 @@ import { useBreadcrumb } from "../frontend/ui/commun/hooks/useBreadcrumb";
 import { RegistrationPage } from "../frontend/ui/registration/RegistrationPage";
 
 export default function PageDInscription() {
-    const { setIsInfoPage } = useContext(BackToSearchContext) as BackToSearchContextValue;
+    const backToSearchContext = useContext(BackToSearchContext) as BackToSearchContextValue;
 
     useBreadcrumb([]);
     useEffect(() => {
-        setIsInfoPage(false);
-        localStorage.clear();
-    }, [])
+        if (backToSearchContext) {
+            backToSearchContext.setIsInfoPage(false);
+            localStorage.clear();
+        }
+    }, [backToSearchContext])
 
     return <RegistrationPage />;
 }
