+++ src/pages/settings/profiles/index.tsx
@@ -14,12 +14,14 @@ type RouterProps = Readonly<{ profiles: ProfilModel[] }>;
 
 export default function Router({ profiles }: RouterProps) {
   const { wording } = useDependencies();
-  const { setIsInfoPage } = useContext(BackToSearchContext) as BackToSearchContextValue;
+  const backToSearchContext = useContext(BackToSearchContext) as BackToSearchContextValue;
 
   useEffect(() => {
-    setIsInfoPage(false);
-    localStorage.clear();
-  }, [])
+    if (backToSearchContext) {
+      backToSearchContext.setIsInfoPage(false);
+      localStorage.clear();
+    }
+  }, [backToSearchContext])
 
   useBreadcrumb([
     {
